requirements.txt
python_libsbml>=5.18.0
clyngor-with-clingo>5.3
clyngor>=0.3.24